requirements.txt
Django>=4.2
djangorestframework
gunicorn
dj-database-url
psycopg2-binary
dotenv